requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
loguru==0.7.0
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
pydantic==1.10.7
python-dotenv==1.0.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.9
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
yarl==1.8.2
gunicorn